requirements.txt
pytest==7.4.3
allure-pytest==2.13.2
selene>=2.0.0rc2
Appium-Python-Client==3.1.1
python-dotenv==1.0.0